wip: continue refactoring.
