+++ fendermint/vm/interpreter/src/fvm/exec.rs
@@ -7,7 +7,7 @@ use super::{
     state::FvmExecState,
     BlockGasLimit, FvmMessage, FvmMessageInterpreter,
 };
-use crate::fvm::activities::{BlockMined, ValidatorActivityTracker};
+use crate::fvm::activity::{BlockMined, ValidatorActivityTracker};
 use crate::ExecInterpreter;
 use anyhow::Context;
 use async_trait::async_trait;
@@ -206,7 +206,7 @@ where
         if let Some(pubkey) = state.block_producer() {
             state
                 .activities_tracker()
-                .track_block_mined(BlockMined { validator: pubkey })?;
+                .record_block_committed(BlockMined { validator: pubkey })?;
         }
 
         let next_gas_market = state.finalize_gas_market()?;
