+++ fendermint/vm/interpreter/src/fvm/state/exec.rs
@@ -3,7 +3,7 @@
 
 use std::collections::{HashMap, HashSet};
 
-use crate::fvm::activities::actor::ActorActivityTracker;
+use crate::fvm::activity::actor::ActorActivityTracker;
 use crate::fvm::externs::FendermintExterns;
 use crate::fvm::gas::BlockGasTracker;
 use anyhow::Ok;
